Start work on spanning tree
TODO:
- make the cache actually work

I want GraphPath.get_path(start_value, end_value, graph_iterator)
to be the main method to GraphPath. It checks to see whether the path
is in the MSP and builds it if it's not, then returns the path

So most of the work (how to interface with the graph, and how to compute
costs) are in the GraphIterator implementations.

I've also added the msp as a list to hold values and make sure they're
not deleted (very important when I translate this to a real man's
language).